api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
feedparser==6.0.10
pymongo==4.6.0
newspaper3k==0.2.8
requests==2.31.0
python-multipart==0.0.6
